ACPI / PPTT: Repainted cache ids from firmware to be less ugly
ACPI added the 'Cache ID' field to the PPTT cache structures to allow
caches to be referred to by other APCI tables. MPAM is an example of
another table that points to caches using this ID. The ID values also
get exposed to user-space via sysfs.

While the cache topology has to be provided to the UEFI firmware as
it can't be detected, the UEFI folk decided to generate the Cache IDs
in such a way that they will be unique. This means they are large
multi-digit numbers, even on a system with a single-digit number of
CPUs. This is less than ideal as these values are exposed to user-space,
and the user has to sometimes type them back in.

As there is no other user of these values, repaint them to be in the
range 0..N. Any helpers that take a cache_id need to specify if it
comes from struct cacheinfo, and if so, unpaint it.

Signed-off-by: James Morse <james.morse@arm.com>
(backported from commit e0abeed173a99aca647813f6db3b69fc676411c7 https://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git/log/?h=mpam/snapshot/v6.12-rc1)
Signed-off-by: KobaK <kobak@nvidia.com>
~~~
Caused by special PPTT table, get the strange id
  $ cat /sys/fs/resctrl/p2/schemata
  MB:196609=100
  L3:196609=fff
Signed-off-by: Tushar Dave <tdave@nvidia.com>
Acked-by: Carol L. Soto <csoto@nvidia.com>
Acked-by: Matthew R. Ochs <mochs@nvidia.com>
Signed-off-by: Matthew R. Ochs <mochs@nvidia.com>